+++ tests/setup.js
@@ -30,6 +30,7 @@ beforeAll(async () => {
       models: process.env.TEST_FOLDER_MODELS || path.resolve('./models'),
       emails:
         process.env.TEST_FOLDER_EMAIL ||
+        process.env.TEST_FOLDER_EMAILS ||
         path.resolve('./services/messaging/email/templates'),
       locales: process.env.TEST_FOLDER_LOCALES || path.resolve('./locales'),
       commands: process.env.TEST_FOLDER_COMMANDS || path.resolve('./commands'),