+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 [BREAKING] change bcrypt encryption with scrypt
 [BREAKING] change internal express parser to formidable parser. Affect you if external formidable is used
 [BREAKING] should not affect any user. Changed email-templates module to internal implementation. Idea to keep dependensy list smaller
+[BREAKING] change i18n middleware to internal one. Nothing should be affected
 [NEW] pagination middleware
+[NEW] requestLogger middleware. Migrated from core server to be an middleware
 [NEW] CreateUser command
 [NEW] custom yup validator for validate File requests
 [UPDATE] updated deps