+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 [NEW] custom yup validator for validate File requests
 [UPDATE] updated deps
 [UPDATE] openApi generator support files
+
+### 3.4.3
+
+[UPDATE] updated deps
+[FIX] fix tests for redis
 [FIX] support in tests TEST_FOLDER_EMAILS
 
 ### 3.4.2